requirements.txt
requests==2.32.3
selenium==4.25.0
Pillow==10.4.0
